# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/items/{id}": {
      "get": {
        "operationId": "ItemController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Item"
        ]
      },
      "put": {
        "operationId": "ItemController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Item"
        ]
      },
      "delete": {
        "operationId": "ItemController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
...
    "/inventory-items": {
      "get": {
        "operationId": "InventoryItemController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "InventoryItem"
        ]
      },
      "post": {
        "operationId": "InventoryItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "InventoryItem"
        ]
      }
    },
    "/inventory-items/{id}": {
      "get": {
        "operationId": "InventoryItemController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "InventoryItem"
        ]
      },
      "put": {
        "operationId": "InventoryItemController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "InventoryItem"
        ]
      },
      "delete": {
        "operationId": "InventoryItemController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "InventoryItem"
        ]
      }
    }
  },
  "info": {
    "title": "LifeRPG API",
    "description": "",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userCharacterId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "desc": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number"
          },
          "amountUnit": {
            "type": "string",
            "enum": [
              "count",
              "minutes"
            ]
          }
        },
        "required": [
          "id",
          "userCharacterId",
          "name",
          "amount",
          "amountUnit"
        ]
      },
...
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "desc": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number"
          },
          "amountUnit": {
            "type": "string",
            "enum": [
              "count",
              "minutes"
            ]
          }
        },
        "required": [
          "name",
          "desc",
          "icon",
          "amount",
          "amountUnit"
        ]
      },
...
      "InventoryItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userCharacterId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "shop",
              "drop",
              "achievement",
              "gift"
            ]
          },
          "acquiredAt": {
            "type": "string"
          },
          "usedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userCharacterId",
          "itemId",
          "source",
          "acquiredAt"
        ]
      },
      "CreateInventoryItemDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "shop",
              "drop",
              "achievement",
              "gift"
            ]
          }
        },
        "required": [
          "itemId"
        ]
      },
      "UpdateInventoryItemDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "shop",
              "drop",
              "achievement",
              "gift"
            ]
          },
          "usedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "itemId",
          "source",
          "usedAt"
        ]
      }
    }
  }
}